Revert "Debug render.com pwd"
This reverts commit fef342ad71b24ebae6c5f6c36c3ba2905315e97e.

+++ build.sh
@@ -2,10 +2,16 @@
 # exit on error
 set -o errexit
 
-echo "...Downloading Chrome"
-mkdir -p ./chrome
-cd ./chrome
-wget -P ./ https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-dpkg -x ./google-chrome-stable_current_amd64.deb ./
-rm ./google-chrome-stable_current_amd64.deb
-cd $HOME/project/src # Make sure we return to where we were
+STORAGE_DIR=/opt/render/project/.render
+
+if [[ ! -d $STORAGE_DIR/chrome ]]; then
+  echo "...Downloading Chrome"
+  mkdir -p $STORAGE_DIR/chrome
+  cd $STORAGE_DIR/chrome
+  wget -P ./ https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+  dpkg -x ./google-chrome-stable_current_amd64.deb $STORAGE_DIR/chrome
+  rm ./google-chrome-stable_current_amd64.deb
+  cd $HOME/project/src # Make sure we return to where we were
+else
+  echo "...Using Chrome from cache"
+fi
